+++ crates/breez-sdk/wasm/src/models/mod.rs
@@ -635,7 +635,7 @@ pub enum ReceivePaymentMethod {
     Bolt11Invoice {
         description: String,
         amount_sats: Option<u64>,
-        expiry_secs: Option<u64>,
+        expiry_secs: Option<u32>,
     },
 }
